DSC stylesheet swapped as test
Switched from dsc6.xsl to one called dsc3nyead.xsl. It is based on solutions that work for us but I recognize it may not be the best solution. It does have some extra things in it that are not essential. It will not include indents to reflect hierarchy of series, subseries, etc. but does eliminate the need to enter lists of values for container type attribute.